minimal changes necessary for run_mrc.py to use text_classification/run_nway_classifier.py instead of run_boolqa_classifier.py
