+++ primeqa/mrc/run_mrc_utils.py
@@ -1,0 +1,35 @@
+import traceback
+from importlib import import_module
+
+def object_reference(reference_as_str: str) -> object:
+    """
+    Given a fully qualified path to a class reference, return a pointer to the reference.
+    This will work with types, functions, methods, and other objects (e.g. dict).
+
+    Args:
+        reference_as_str: the fully qualified path (expects the fully qualified path in dot notation,
+                          e.g. primeqa.mrc.processors.postprocessors.extractive.ExtractivePostProcessor).
+
+    Returns:
+        reference to path given by input
+
+    Raises:
+        TypeError: Unable to resolve input path
+    """
+    def _split_into_class_and_module_name(class_path):
+        modules = class_path.split('.')
+        if len(modules) > 1:
+            return ".".join(modules[:-1]), modules[-1]
+        else:
+            return class_path, None
+
+    try:
+        module_name, object_name = _split_into_class_and_module_name(reference_as_str)
+        module_reference = import_module(module_name)
+        if object_name is None:
+            return module_reference
+        else:
+            return getattr(module_reference, object_name)
+    except Exception as ex:
+        traceback.print_exc()  # Shows additional traceback for why imports fail
+        raise TypeError(f"Unable to resolve the string {reference_as_str} to a fully qualified class path") from ex
